+++ library/extractor/src/main/java/com/google/android/exoplayer2/extractor/mp4/Atom.java
@@ -281,6 +281,9 @@ import java.util.List;
   public static final int TYPE_TTML = 0x54544d4c;
 
   @SuppressWarnings("ConstantCaseForConstants")
+  public static final int TYPE_m1v_ = 0x6d317620;
+
+  @SuppressWarnings("ConstantCaseForConstants")
   public static final int TYPE_mp4v = 0x6d703476;
 
   @SuppressWarnings("ConstantCaseForConstants")